pyproject.toml
[project]
name = "sample-backend-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dj-database-url>=3.0.1",
    "django-cors-headers>=4.7.0",
    "django-storages>=1.14.6",
    "djangorestframework>=3.16.0",
    "djoser>=2.3.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.35.0",
    "whitenoise>=6.9.0",
]
